refactor(Repository): Dao 삭제, Dao 코드 Repository로 이동
