+++ src/main/java/roomescape/theme/ThemeRepository.java
@@ -1,37 +1,102 @@
 package roomescape.theme;
 
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 
+import java.sql.PreparedStatement;
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Repository
 public class ThemeRepository {
 
-    private final ThemeDao themeDao;
+    private final JdbcTemplate jdbcTemplate;
+    private final RowMapper<Theme> rowMapper = (rs, rowNum) -> new Theme(
+            rs.getLong("id"),
+            rs.getString("name"),
+            rs.getString("description"),
+            rs.getString("thumbnail")
+    );
 
-    public ThemeRepository(ThemeDao themeDao) {
-        this.themeDao = themeDao;
+    public ThemeRepository(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
     }
 
     public Theme save(String name, String description, String thumbnail) {
-        return themeDao.save(name, description, thumbnail);
+        String sql = "INSERT INTO themes (name, description, thumbnail) VALUES (?, ?, ?)";
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+
+        jdbcTemplate.update(con -> {
+            PreparedStatement ps = con.prepareStatement(sql, new String[]{"id"});
+            ps.setString(1, name);
+            ps.setString(2, description);
+            ps.setString(3, thumbnail);
+            return ps;
+        }, keyHolder);
+
+        long id = Objects.requireNonNull(keyHolder.getKey()).longValue();
+        return new Theme(id, name, description, thumbnail);
     }
 
     public void delete(long id) {
-        themeDao.delete(id);
-    }
+        String sql = "DELETE FROM themes WHERE id = ?";
 
-    public List<Theme> findRanked(String sort, String order, LocalDate startDate, LocalDate endDate, Long limit) {
-        return themeDao.findRanked(sort, order, startDate, endDate, limit);
+        jdbcTemplate.update(sql, id);
     }
 
     public List<Theme> findAll() {
-        return themeDao.findAll();
+        String sql = "SELECT * FROM themes";
+        return jdbcTemplate.query(sql, rowMapper);
+    }
+
+    public List<Theme> findRanked(String sort, String order, LocalDate startDate, LocalDate endDate, Long limit) {
+        String sql = getReservationSortSql(sort, order, limit);
+        return jdbcTemplate.query(sql, rowMapper, startDate, endDate);
+    }
+
+    private String getReservationSortSql(String sort, String order, Long limit) {
+        String sortColumn = "reservationCount";
+        if ("id".equalsIgnoreCase(sort)) {
+            sortColumn = "id";
+        } else if ("name".equalsIgnoreCase(sort)) {
+            sortColumn = "name";
+        }
+
+        String sortOrder = "DESC";
+        if ("ASC".equalsIgnoreCase(order)) {
+            sortOrder = "ASC";
+        }
+
+        StringBuilder sql = new StringBuilder(
+                "SELECT t.id, t.name, t.description, t.thumbnail, COUNT(r.id) AS reservationCount " +
+                        "FROM themes t " +
+                        "INNER JOIN reservation r ON t.id = r.theme_id " +
+                        "WHERE r.date >= ? AND r.date <= ? " +
+                        "GROUP BY t.id, t.name, t.description, t.thumbnail " +
+                        "ORDER BY " + sortColumn + " " + sortOrder
+        );
+
+        if (limit != null) {
+            sql.append(" LIMIT ").append(limit);
+        }
+
+        return sql.toString();
     }
 
     public Optional<Theme> findById(long id) {
-        return themeDao.findById(id);
+        String sql = "SELECT * FROM themes WHERE id = ?";
+
+        try {
+            Theme theme = jdbcTemplate.queryForObject(sql, rowMapper, id);
+            return Optional.ofNullable(theme);
+        } catch (EmptyResultDataAccessException exception) {
+            return Optional.empty();
+        }
     }
 }
